+++ src/test/java/com/luq/store/repositories/ProductRepositoryTest.java
@@ -12,6 +12,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.data.domain.Sort;
 import org.springframework.test.annotation.DirtiesContext;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -44,11 +45,11 @@ public class ProductRepositoryTest {
 
         fakeProduct1 = pRepository.save(new Product(
             "XOneCont", "Xbox One Controller", "Controller for Xbox One Console",
-            fakeSupplier1, 200.00F
+            fakeSupplier1, BigDecimal.valueOf(200.00)
         ));
         fakeProduct2 = pRepository.save(new Product(
             "PS5Cont", "Playstation 5 Controller", "Controller for Playstation 5 Console",
-            fakeSupplier2, 250.00F
+            fakeSupplier2, BigDecimal.valueOf(250.00)
         ));
     }
